Fixed data type and added test for time

+++ swmmtonetcdf/swmmtonetcdf.py
@@ -128,7 +128,7 @@ def create_netcdf_from_swmm(swmm_output_file: str, netcdf_output_file: str, read
     netcdf_output.createDimension(dimname='time', size=None)
     nc_time_variable = netcdf_output.createVariable(
         varname="time",
-        datatype=np.float32,
+        datatype=np.float,
         dimensions=("time",),
 
     )
@@ -218,25 +218,25 @@ def create_netcdf_from_swmm(swmm_output_file: str, netcdf_output_file: str, read
 
     nc_node_timeseries = netcdf_output.createVariable(
         varname='node_timeseries',
-        datatype=np.float32,
+        datatype=np.float,
         dimensions=('nodes', 'node_attributes', 'time',)
     )
 
     nc_link_timeseries = netcdf_output.createVariable(
         varname='link_timeseries',
-        datatype=np.float32,
+        datatype=np.float,
         dimensions=('links', 'link_attributes', 'time',)
     )
 
     nc_catchment_timeseries = netcdf_output.createVariable(
         varname='catchment_timeseries',
-        datatype=np.float32,
+        datatype=np.float,
         dimensions=('catchments', 'catchment_attributes', 'time',)
     )
 
     nc_system_timeseries = netcdf_output.createVariable(
         varname='system_timeseries',
-        datatype=np.float32,
+        datatype=np.float,
         dimensions=('system_attributes', 'time',)
     )
 
@@ -290,7 +290,7 @@ def create_netcdf_from_swmm(swmm_output_file: str, netcdf_output_file: str, read
                     endPeriod=num_steps
                 )
 
-                nc_node_timeseries[j, i, :] = np.array(node_series)
+                nc_node_timeseries[j, i, :] = np.array(node_series, dtype=np.float)
                 netcdf_output.sync()
                 j += 1
 
@@ -309,7 +309,7 @@ def create_netcdf_from_swmm(swmm_output_file: str, netcdf_output_file: str, read
                     endPeriod=num_steps
                 )
 
-                nc_link_timeseries[j, i, :] = np.array(link_series)
+                nc_link_timeseries[j, i, :] = np.array(link_series, dtype=np.float)
                 netcdf_output.sync()
                 j += 1
 
@@ -332,21 +332,22 @@ def create_netcdf_from_swmm(swmm_output_file: str, netcdf_output_file: str, read
             # catchment attributes
             for j in range(num_catchments):
                 catchment_results = output.get_subcatch_result(p_handle=file_handle, timeIndex=t, subcatchIndex=j)
-                nc_catchment_timeseries[j, :, t] = np.array(catchment_results[0:num_catchment_attributes])
+                nc_catchment_timeseries[j, :, t] = np.array(catchment_results[0:num_catchment_attributes],
+                                                            dtype=np.float)
 
             # node attributes
             for j in range(num_nodes):
                 node_results = output.get_node_result(p_handle=file_handle, timeIndex=t, nodeIndex=j)
-                nc_node_timeseries[j, :, t] = np.array(node_results[0:num_node_attributes])
+                nc_node_timeseries[j, :, t] = np.array(node_results[0:num_node_attributes], dtype=np.float)
 
             # link attributes
             for j in range(num_links):
                 link_results = output.get_link_result(p_handle=file_handle, timeIndex=t, linkIndex=j)
-                nc_link_timeseries[j, :, t] = np.array(link_results[0:num_link_attributes])
+                nc_link_timeseries[j, :, t] = np.array(link_results[0:num_link_attributes], dtype=np.float)
 
             # system attributes
             system_results = output.get_system_result(p_handle=file_handle, timeIndex=t, dummyIndex=0)
-            nc_system_timeseries[:, t] = np.array(system_results[0:num_system_attributes])
+            nc_system_timeseries[:, t] = np.array(system_results[0:num_system_attributes], dtype=np.float)
 
             if t % 5000 == 0:
                 netcdf_output.sync()
